Crate eosio_macros_internal

Internal use only.

This crate is an implementation detail that will hopefully go away once the proc_macro_hygiene feature is stabilized. In the meantime we must use this crate (and proc_macro_hack) to allow for function-like procedural macros in expression positions.

Macros

abi

Attribute Macros

action

table

Derive Macros

NumBytes
Derive the NumBytes trait

Read
Derive the Read trait

Table
TODO docs

Write
Derive the Write trait

proc_macro_hack_n

proc_macro_hack_s